Develop

+++ src/main/java/jp/co/waja/core/service/team/TeamService.java
@@ -1,16 +1,16 @@
 package jp.co.waja.core.service.team;
 
 import jp.co.waja.core.entity.Team;
-import jp.co.waja.core.model.team.TeamCreateRequest;
-import jp.co.waja.core.model.team.TeamEditRequest;
+import jp.co.waja.core.model.team.*;
+import jp.co.waja.core.repository.staff.StaffRepository;
 import jp.co.waja.core.repository.team.TeamRepository;
-import jp.co.waja.exception.NotFoundException;
+import jp.co.waja.exception.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 @Transactional
@@ -19,14 +19,18 @@ public class TeamService {
 	@Autowired
 	private TeamRepository teamRepository;
 
-	public List<Team> teams() {
+	@Autowired
+	private StaffRepository staffRepository;
+
+	public List<Team> getTeams() {
 		return teamRepository.findAll();
 	}
 
-	public Team findOneById(long teamId) {
+	public Team getTeam(long teamId) {
 		return teamRepository.findOneById(teamId);
 	}
 
+	@PreAuthorize("hasRole('ADMIN')")
 	public Team create(TeamCreateRequest request) {
 		Team team = new Team();
 		team.setName(request.getName());
@@ -34,6 +38,7 @@ public class TeamService {
 		return teamRepository.saveAndFlush(team);
 	}
 
+	@PreAuthorize("hasRole('ADMIN')")
 	public Team edit(TeamEditRequest request, Long id) {
 		Optional<Team> teamOptional = Optional.ofNullable(teamRepository.findOneById(id));
 		if (!teamOptional.isPresent()) {
@@ -46,8 +51,19 @@ public class TeamService {
 		return teamRepository.saveAndFlush(team);
 	}
 
-	public void delete(Long id) {
-		// Teamに１人でもStaffが所属していたら削除しない
+	@PreAuthorize("hasRole('ADMIN')")
+	public Optional<String> delete(Long id) throws NotFoundException, WrongDeleteException {
+		Team team = teamRepository.findOneById(id);
+		if (Objects.isNull(team)) {
+			throw new NotFoundException("team");
+		}
+
+		long staffCount = staffRepository.countByTeam(team);
+		if (staffCount > 0) {
+			throw new WrongDeleteException("StaffExist");
+		}
+
 		teamRepository.delete(id);
+		return Optional.ofNullable(team.getName());
 	}
 }
